requirements.txt
flask==2.0.1
python-dotenv==0.19.0
mysql-connector-python==8.0.26
requests==2.26.0
python-jose==3.3.0
pandas
google-cloud-aiplatform
slowapi 